refactor: rely on typed messages for communication with vscode-ui
