+++ libs/vscode-ui/feature-task-execution-form/src/index.ts
@@ -1,4 +1,3 @@
 export { TaskExecutionFormComponent } from './lib/task-execution-form.component';
 
 export * from './lib/vscode-ui-feature-task-execution-form.module';
-export * from './lib/task-execution-form.schema';
