--- libs/vscode-ui/feature-task-execution-form/src/lib/task-execution-form.schema.ts
@@ -1,5 +1,0 @@
-import { InjectionToken } from '@angular/core';
-
-export const TASK_EXECUTION_SCHEMA = new InjectionToken(
-  'VSCODE_UI_FEATURE_TASK_EXECUTION_FORM_SCHEMA'
-);
