+++ libs/vscode-ui/feature-task-execution-form/src/lib/vscode-ui-feature-task-execution-form.module.ts
@@ -7,19 +7,6 @@ import { FormatTaskPipe } from './format-task/format-task.pipe';
 import { VscodeUiComponentsModule } from '@nx-console/vscode-ui/components';
 import { ArgumentListModule } from '@nx-console/vscode-ui/argument-list';
 
-import { TASK_EXECUTION_SCHEMA } from './task-execution-form.schema';
-import { TaskExecutionSchema } from '@nx-console/shared/schema';
-
-declare global {
-  interface Window {
-    VSCODE_UI_SCHEMA: TaskExecutionSchema;
-  }
-}
-
-export function getSchema() {
-  return window.VSCODE_UI_SCHEMA;
-}
-
 @NgModule({
   imports: [
     CommonModule,
@@ -30,11 +17,5 @@ export function getSchema() {
   ],
   declarations: [TaskExecutionFormComponent, FormatTaskPipe],
   exports: [TaskExecutionFormComponent],
-  providers: [
-    {
-      provide: TASK_EXECUTION_SCHEMA,
-      useFactory: getSchema,
-    },
-  ],
 })
 export class VscodeUiFeatureTaskExecutionFormModule {}
